Add instructions to run migrate on first run
Currently documentation overall is very confusing. Honestly I'd rewrite it completely, but the least that can be done is to at least add an instruction to run the `shiori migrate` command on first run. I just spent almost an hour digging through issues to find this out. I'm certain that if I were to install this again, I would love to see this instruction here.